+++ src/test/resources/application.yml
@@ -188,6 +188,7 @@ uk:
             SOL_PERSONAL_SERVICE: 1e2f2ac5-ce4c-4fce-aca0-47a876eda089
             SOL_DN_DECISION_MADE: fd6d9667-527f-4c22-b1f1-d095c412ab2c
             SOL_DA_GRANTED_NOTIFICATION: fc8f7343-8f07-42b7-a3c5-52543a21015d
+            CITIZEN_DEEMED_APPROVED: 00f27db6-2678-4ccd-8cdd-44971b330ca4
           welsh:
             AOS_RECEIVED_NO_ADMIT_ADULTERY: e0791556-a202-45e5-9b55-7d1bb61415db
             AOS_RECEIVED_NO_ADMIT_ADULTERY_CORESP_NOT_REPLIED: bf8945dc-f6e5-48fa-adb6-5f277441584c
@@ -237,6 +238,7 @@ uk:
             SOL_PERSONAL_SERVICE: 1e2f2ac5-ce4c-4fce-aca0-47a876eda089
             SOL_DN_DECISION_MADE: fd6d9667-527f-4c22-b1f1-d095c412ab2c
             SOL_DA_GRANTED_NOTIFICATION: fc8f7343-8f07-42b7-a3c5-52543a21015d
+            CITIZEN_DEEMED_APPROVED: 00f27db6-2678-4ccd-8cdd-44971b330ca4
 
 court-distribution-config:
   distribution:
